+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+Version 0.21
+------------
+
 Version 0.20
 ------------
 - `dlurlserver()` now returns NULL for no server